Refactor RefCell logic into wrapper type
